Flag app as isMasqueradable for accountId setting
For some reason I haven't uncovered yet, our SmartPBX app is loading up
with isMasqueradable = false, despite the core _init_app function
indicating that it should take the property of the app within the
monster.apps.appStore[app.name] property, which is `true` in our case.

So as a hack, I am just statically declaring this property, which I
hope will cause the accountId property to set when the app is
initialized.